little bug fix with size types

+++ ns-3.16/src/mpi/model/distributed-simulator-impl.h
@@ -89,7 +89,7 @@ struct parmetis_graph_t
 {
 
   // num of vertex (global), num of vertex (local)
-  int gnvtxs, nvtxs;
+  size_t gnvtxs, nvtxs;
 
   // physical vertex distribution (for partition computing)
   parmetis_idx_t* vtxdist;
